% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpl.R
\name{bpl_search}
\alias{bpl_search}
\title{Search Brooklyn Public Library branches by keyword}
\usage{
bpl_search(query)
}
\arguments{
\item{query}{Character string to search for (case-insensitive)}
}
\value{
tibble: same columns as bpl_branches()
}
\description{
Search Brooklyn Public Library branches by keyword
}
