% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpl.R
\name{bpl_search_resources}
\alias{bpl_search_resources}
\title{Search Brooklyn Public Library electronic resources}
\usage{
bpl_search_resources(query)
}
\arguments{
\item{query}{Character string to search for (case-insensitive)}
}
\value{
tibble: title, description, link
}
\description{
Search Brooklyn Public Library electronic resources
}
